% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_surv.R
\name{forest_surv}
\alias{forest_surv}
\title{Forest plot for meta-analysis of survival rates}
\usage{
forest_surv(
  metares,
  method = "log-log",
  percent = TRUE,
  xlim = NULL,
  estlab = "Survival (\%)",
  digits = 1,
  n = "n",
  ...
)
}
\arguments{
\item{metares}{a \code{metagen} object for the meta-analysis of survival
rate.}

\item{method}{a character value specifying the estimation method for
confidence intervals used in the \code{impute_se_surv} function.
Choose from "plain", "log", and "log-log". Default is "log-log".}

\item{percent}{a logical value that specifies the scale of the display;
\code{TRUE} for a percentage scale (0 to 100),
\code{FALSE} for a rate scale (0 to 1).
Default is \code{TRUE}.}

\item{xlim}{a numeric vector of length 2 that specifies the horizontal limits
of the plot region. If unspecified, the function sets the horizontal plot
limits to some sensible values.}

\item{estlab}{a character value for the label of the estimation target.
Default is "Survival (\%)".}

\item{digits}{a numeric value specifying the number of decimal places to
be displayed. Default is \code{1}.}

\item{n}{a character value for the Variable name for the sample size of
each study. Default is “n”.}

\item{...}{Additional arguments to be passed to the \code{forest} function.}
}
\value{
a forest plot created with the \code{forest} function in the
  \code{meta} package.
}
\description{
\code{forest_surv} is a function to draw a forest plot using a
\code{metagen} object of a survival rate meta-analysis based on the
\code{metagen} function.
The data frame output from the \code{impute_se_surv} function
should be used as input to the \code{metagen} function.
}
\details{
In meta-analyses of survival rates, precision information (i.e.,
confidence intervals) may be missing in some studies.
Excluding studies with missing precision information may significantly reduce
the accuracy and reliability of survival rate estimators in meta-analyses.
This function imputes precision information using data commonly
available from study literature, such as sample size, number of events,
and risk set size at a time point of interest.
It then returns a data frame with the variables necessary for meta-analysis
added. Even when no data are missing, it can still be used as a
preprocessing step in meta-analysis of survival rates.
}
\examples{
library(metaSurvMissCI)
library(meta)
data("metadata_chordoma")
impdata <- impute_se_surv(data = metadata_chordoma, St = "PFS5y",
                       LCL = "PFSL5y", UCL = "PFSU5y", n = "n",
                       nt = "n_5yPFS", ne = "ne_PFS", p = "pr_PFS")
meta_imp <- metagen(TE = tr_St, seTE = tr_SE, studlab = Study,
                    data = impdata, method.tau = "REML")
forest_surv(meta_imp, xlim = c(30, 100), estlab = "5-year PFS (\%)")

}
\seealso{
\code{\link[meta]{metagen}}
}
